model/requirements.txt
tensorflow
pulsar-client
scikit-learn
psycopg2-binary
pandas
numpy
matplotlib
joblib
boto3
pulsar-client[avro]
